requirements.txt
aiolimiter==1.1.0
anyio==4.0.0
APScheduler==3.10.4
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
cryptography==41.0.4
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
python-telegram-bot==20.5
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
tornado==6.3.3
tzlocal==5.0.1
